requirements.txt
langchain
openai==0.27.8
python-dotenv==1.0.0
google-search-results==2.4.2
duckduckgo-search==3.8.5
diskcache==5.6.3
PyMuPDF==1.23.3
bs4==0.0.1
pypdf==3.16.0
tabulate==0.9.0
tldextract==3.5.0
faiss-cpu
Chroma==0.2.0
tiktoken==0.5.0
selenium==4.12.0
unstructured==0.10.14
matplotlib==3.8.0
matplotlib_venn==0.11.9
chromadb==0.4.12
seaborn==0.12.2
tabulate==0.9.0
streamlit==1.26.0
pandas==2.1.1
Markdown==3.4.4
sentence-transformers
unstructured
ipython==8.15.0
tabula-py==2.8.2
llama-hub==0.0.43
llama-index==0.8.61
braintrust==0.0.64
autoevals==0.0.28
transformers==4.35.0
torch==2.1.0
torchvision==0.16.0
pdfkit==1.0.0







